+++ core-ng/src/main/java/core/framework/internal/db/RepositoryImpl.java
@@ -83,47 +83,45 @@ public final class RepositoryImpl<T> implements Repository<T> {
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, insertedRows);
-            logger.debug("insertIgnore, sql={}, params={}, insertedRows={}, elapsed={}", sql, new SQLParams(database.operation.enumMapper, params), insertedRows, elapsed);
+            logger.debug("insertIgnore, sql={}, params={}, inserted={}, elapsed={}", sql, new SQLParams(database.operation.enumMapper, params), insertedRows == 1, elapsed);
             database.checkOperation(elapsed, operations);
         }
     }
 
     @Override
-    public void upsert(T entity) {
+    public boolean upsert(T entity) {
         var watch = new StopWatch();
         if (insertQuery.generatedColumn != null) throw new Error("entity must not have auto increment primary key, entityClass=" + entityClass.getCanonicalName());
         validator.validate(entity, false);
-        int updatedRows = 0;
+        int affectedRows = 0;
         String sql = insertQuery.upsertSQL();
         Object[] params = insertQuery.params(entity);
         try {
-            int result = database.operation.update(sql, params);
+            affectedRows = database.operation.update(sql, params);
             // refer to https://dev.mysql.com/doc/refman/8.0/en/insert-on-duplicate.html
             // With ON DUPLICATE KEY UPDATE, the affected-rows value per row is 1 if the row is inserted as a new row, 2 if an existing row is updated, and 0 if an existing row is set to its current values.
-            if (result > 0) updatedRows = 1;
+            return affectedRows == 1;
         } finally {
             long elapsed = watch.elapsed();
-            int operations = ActionLogContext.track("db", elapsed, 0, updatedRows);
-            logger.debug("upsert, sql={}, params={}, updatedRows={}, elapsed={}", sql, new SQLParams(database.operation.enumMapper, params), updatedRows, elapsed);
+            int operations = ActionLogContext.track("db", elapsed, 0, affectedRows == 0 ? 0 : 1);
+            logger.debug("upsert, sql={}, params={}, inserted={}, elapsed={}", sql, new SQLParams(database.operation.enumMapper, params), affectedRows == 1, elapsed);
             database.checkOperation(elapsed, operations);
         }
     }
 
     @Override
     public boolean update(T entity) {
-        int updatedRows = update(entity, false, null, null);
-        if (updatedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
-        return updatedRows == 1;
+        return update(entity, false, null, null);
     }
 
-    private int update(T entity, boolean partial, String where, Object[] params) {
+    private boolean update(T entity, boolean partial, String where, Object[] params) {
         var watch = new StopWatch();
         validator.validate(entity, partial);
         UpdateQuery.Statement query = updateQuery.update(entity, partial, where, params);
         int updatedRows = 0;
         try {
             updatedRows = database.operation.update(query.sql, query.params);
-            return updatedRows;
+            return updatedRows == 1;
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, updatedRows);
@@ -134,15 +132,12 @@ public final class RepositoryImpl<T> implements Repository<T> {
 
     @Override
     public boolean partialUpdate(T entity) {
-        int updatedRows = update(entity, true, null, null);
-        if (updatedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
-        return updatedRows == 1;
+        return update(entity, true, null, null);
     }
 
     @Override
     public boolean partialUpdate(T entity, String where, Object... params) {
-        int updatedRows = update(entity, true, where, params);
-        return updatedRows == 1;
+        return update(entity, true, where, params);
     }
 
     @Override
@@ -151,14 +146,14 @@ public final class RepositoryImpl<T> implements Repository<T> {
         if (primaryKeys.length != selectQuery.primaryKeyColumns) {
             throw new Error(Strings.format("the length of primary keys must match columns, primaryKeys={}, columns={}", primaryKeys.length, selectQuery.primaryKeyColumns));
         }
-        int deletedRows = 0;
+        int affectedRows = 0;
         try {
-            deletedRows = database.operation.update(deleteSQL, primaryKeys);
-            if (deletedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", deletedRows);
-            return deletedRows == 1;
+            affectedRows = database.operation.update(deleteSQL, primaryKeys);
+            if (affectedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", affectedRows);
+            return affectedRows == 1;
         } finally {
             long elapsed = watch.elapsed();
-            int operations = ActionLogContext.track("db", elapsed, 0, deletedRows);
+            int operations = ActionLogContext.track("db", elapsed, 0, affectedRows);
             logger.debug("delete, sql={}, params={}, elapsed={}", deleteSQL, new SQLParams(database.operation.enumMapper, primaryKeys), elapsed);
             database.checkOperation(elapsed, operations);
         }
@@ -198,10 +193,10 @@ public final class RepositoryImpl<T> implements Repository<T> {
         }
         int insertedRows = 0;
         try {
-            int[] results = database.operation.batchUpdate(sql, params);
-            boolean[] insertedResults = new boolean[results.length];
-            insertedRows = updatedResults(results, insertedResults);
-            return insertedResults;
+            int[] affectedRows = database.operation.batchUpdate(sql, params);
+            boolean[] results = new boolean[affectedRows.length];
+            insertedRows = batchResults(affectedRows, results);
+            return results;
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, insertedRows);
@@ -211,7 +206,7 @@ public final class RepositoryImpl<T> implements Repository<T> {
     }
 
     @Override
-    public void batchUpsert(List<T> entities) {
+    public boolean[] batchUpsert(List<T> entities) {
         var watch = new StopWatch();
         if (entities.isEmpty()) throw new Error("entities must not be empty");
         if (insertQuery.generatedColumn != null) throw new Error("entity must not have auto increment primary key, entityClass=" + entityClass.getCanonicalName());
@@ -223,8 +218,10 @@ public final class RepositoryImpl<T> implements Repository<T> {
         }
         int updatedRows = 0;
         try {
-            int[] results = database.operation.batchUpdate(sql, params);
-            updatedRows = database.batchUpdatedRows(results);
+            int[] affectedRows = database.operation.batchUpdate(sql, params);
+            boolean[] results = new boolean[affectedRows.length];
+            updatedRows = batchResults(affectedRows, results);
+            return results;
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, updatedRows);
@@ -247,11 +244,11 @@ public final class RepositoryImpl<T> implements Repository<T> {
         }
         int deletedRows = 0;
         try {
-            int[] results = database.operation.batchUpdate(deleteSQL, params);
-            boolean[] deletedResults = new boolean[results.length];
-            deletedRows = updatedResults(results, deletedResults);
-            if (deletedRows != primaryKeys.size()) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows do not match size of primary keys, results={}", Arrays.toString(results));
-            return deletedResults;
+            int[] affectedRows = database.operation.batchUpdate(deleteSQL, params);
+            boolean[] results = new boolean[affectedRows.length];
+            deletedRows = batchResults(affectedRows, results);
+            if (deletedRows != primaryKeys.size()) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows do not match size of primary keys, results={}", Arrays.toString(affectedRows));
+            return results;
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, deletedRows);
@@ -262,14 +259,17 @@ public final class RepositoryImpl<T> implements Repository<T> {
 
     // use procedurally way to produce results to insertedRows and insertedResults in one method, to balance both performance and ease of unit test
     // not recommended applying in application level code
-    int updatedResults(int[] results, boolean[] updatedResults) {
+    int batchResults(int[] affectedRows, boolean[] results) {
         int updatedRows = 0;
-        for (int i = 0; i < results.length; i++) {
-            int result = results[i];
+        for (int i = 0; i < affectedRows.length; i++) {
+            int affectedRow = affectedRows[i];
             // with batchInsert, mysql returns -2 if insert succeeds, for batch only
             // refer to com.mysql.cj.jdbc.ClientPreparedStatement.executeBatchedInserts
-            if (result == Statement.SUCCESS_NO_INFO || result > 0) {
-                updatedResults[i] = true;
+            // with upsert, returns 1 if inserted, 2 if updated
+            if (affectedRow == Statement.SUCCESS_NO_INFO || affectedRow == 1) {
+                results[i] = true;  // return true if row is inserted/deleted/updated
+                updatedRows++;
+            } else if (affectedRow > 0) {
                 updatedRows++;
             }
         }